pyproject.toml
[tool.poetry]
name = "tafker"
version = "0.1.0"
description = "Simple automation system for the Linux Desktop"
authors = ["Philipp Schmitt <philipp@schmitt.co>"]
license = "GPL-3"

[tool.poetry.dependencies]
python = "^3.9"
rich = "^10.12.0"
psutil = "^5.8.0"
asyncclick = "^8.0.1"
anyio = "^3.3.4"
uvloop = "^0.16.0"
PyYAML = "^6.0"
xdg = "^5.1.1"
python-xlib = "^0.31"
ewmh = "^0.1.6"
setproctitle = "^1.2.2"

[tool.poetry.dev-dependencies]
ipython = "^7.28.0"
black = "^21.9b0"
isort = "^5.9.3"


[tool.poetry.scripts]
tafker = "tafker.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
